creating a single package with only the common part for ROS transports
NOTE: this is only meant to have a singled-out package out of the
rtt_ros_integration for ROS transports, for use later in Rock.

It is highly experimental AND bad development practice, as I did not (yet)
synchronize with the original writers of the ROS-package rtt_ros_integration.
This will be done before the ROS integration in Rock is fully released.